Adds fields supporting no-sandbox mode to IR module. (#10)
* Adds fields supporting no-sandbox mode to IR module.

This change adds fields that represent relocation information for
function- and data pointers (both within code as well as within
static initializers) and mappings from pointers to relevant indices.

Relocation info can be 32-bit or 64-bit, but depending on the selected
memory model only one of these sets should ever be non-empty.

There are two pointer-to-index mappings:

Function pointers:  WASM-style function pointers (integer constants
that refer to entries in the indirect function table) are mapped to
the corresponding index of the function in question.  The index lets
us find the relevant entry in module_->funcs.

Data pointers:  WASM-style data pointers pointing at the LAST byte of
a piece of global data are mapped to the corresponding index of the
DataSegment in question, making it possible to find the relevant entry
in module_->data_segments.
By using the last byte it becomes possible to use C++
std::map::lower_bound on any address within the range of addresses
of a given data segment in order to obtain the index of that segment.

binary-reader-ir.cc is modified to populate these fields accordingly.
The offset of element- and data segments is assumed to be given by a
single ixx.const instruction.

* Uses symbol table information to interpret reloc info.

* Unconditionally store data segment index.

* Moves validity check of symbol up.

This rearranges some of the logic in On{Function,Data}Symbol so that
the function index or the data segment index gets validated before
being stored in the symbol->index map.