No sandbox data segments (#22)
* Adds fields supporting no-sandbox mode to IR module.

This change adds fields that represent relocation information for
function- and data pointers (both within code as well as within
static initializers) and mappings from pointers to relevant indices.

Relocation info can be 32-bit or 64-bit, but depending on the selected
memory model only one of these sets should ever be non-empty.

There are two pointer-to-index mappings:

Function pointers:  WASM-style function pointers (integer constants
that refer to entries in the indirect function table) are mapped to
the corresponding index of the function in question.  The index lets
us find the relevant entry in module_->funcs.

Data pointers:  WASM-style data pointers pointing at the LAST byte of
a piece of global data are mapped to the corresponding index of the
DataSegment in question, making it possible to find the relevant entry
in module_->data_segments.
By using the last byte it becomes possible to use C++
std::map::lower_bound on any address within the range of addresses
of a given data segment in order to obtain the index of that segment.

binary-reader-ir.cc is modified to populate these fields accordingly.
The offset of element- and data segments is assumed to be given by a
single ixx.const instruction.

* No-sandbox function pointer creation.

* Uses symbol table information to interpret reloc info.

* Use IR maps properly for function pointer translation.

* Unconditionally store data segment index.

* Moves validity check of symbol up.

This rearranges some of the logic in On{Function,Data}Symbol so that
the function index or the data segment index gets validated before
being stored in the symbol->index map.

* Implements no_sandbox mode for ixx.const.

Uses relocation info to render integer constants either as the
function pointers or the memory pointers that they represent.

* Un-sandboxing memory loads.

* Factor function and data reloc handling into methods.

* No-sandboxify load instruction.

* Prefix arg with default for data address writer.

* Avoid writing 0 offset in no-sandbox load instruction.

* No-sandboxify store instruction.

* Changes how reloc data is represented in IR.

The reloc mappings are now pointing at symbols (as they do in the
binary itself) rather than being pre-composed with the mappings from
symbols to functions and data segments.  The reason is that for
imported data there is no corresponding data segment, so pre-composition
does not work there.

Instead, undefined data symbols are separately stored by name.

* Handle undefined data symbols.

Uses the new IR reloc data.

* Reintroduce index validation.

* Fix pointer bug in handling of memory address relocation.

* Some refactoring of address mechanics.

* Fix no-sandbox logic in ConstExpr writer.

* Addressed review comments.

* Initial cut for writing no-sandbox data segments.

Initializers have ot be fleshed out.

There is currently a problem with data offsets.
Relocation info is not correct.  This has
to be fixed separately.

* Memory address mapping.

* Recording data segment address mapping.

* Use data segment address from recorded mapping.

* Use memory offsets in data init maps.

* Mapping by last byte to make lower_bound work.

* Added data offset to DataSegment.

* Data segment defs now working.

* Undo spurious formatting.

* Patch Write for SimdShuffleOpExpr.

* Patch Write for LoadSplatExpr from main.

* Temporarily back out of change to LoadSplatExpr.

* Addressed formatting suggestions from review.

* Apply memory transform to other load/store instructions.

* Fixed formatting typo.

* Data segment initialization - first cut!

Seems to work as intended.

* Some refactoring to share code.

* Formatted.

* Back out of cosmetic change.

Don't use WriteBytes in place of the original code that
writes data segment bytes.  The formatting is slightly different
(WriteBytes is arguably nicer), but it upsets all the tests.

* Generate data forward-declarations.

Also, data-segments must be non-const since we are now using
them directly as data (which in general is mutable).

* Fix missing write of prefix.

* Minor cosmetic fix.

* Use C cast, not reinterpret_cast.