Big simplification of data init code. (#40)
The original implementation needlessly converted WASM file
offsets to memory addresses and then worked in that space.
It turns out that one can equivalently work directly within
the file offset space (or rather: the space of offsets relative
to the start of the data section within that file).

This greatly simplifies the code
because no conversion to memory addresses is needed anywhere,
eliminating data structures (the mapping from data segments to
memory addresses) as well as removing all the code used to
maintain these data structures.

The result is also more robust: Conversion to memory addresses
had to obtain the base address for each memory segment, and that
base address can in general be computed by a sequence of WASM
instructions.  The conversion code had to make assumptions about
those instructions (namely that there is exactly one, and that it
is an i{32,64}.const).  These assumptions are currently satisfied,
but it is better if we don't have to rely on them.